Start reorganizing CUDA.
