+++ include/aluminum/cuda/CMakeLists.txt
@@ -1,0 +1,10 @@
+set_source_path(THIS_DIR_CUDA_HEADERS
+  cuda.hpp
+  helper_kernels.hpp
+  )
+
+if (AL_HAS_CUDA)
+  list(APPEND THIS_DIR_HEADERS "${THIS_DIR_CUDA_HEADERS}")
+endif ()
+
+set(ALUMINUM_HEADERS "${HEADERS}" "${THIS_DIR_HEADERS}" PARENT_SCOPE)
