+++ include/aluminum/ht/base_state.hpp
@@ -28,7 +28,7 @@
 #pragma once
 
 #include "aluminum/progress.hpp"
-#include "aluminum/cuda.hpp"
+#include "aluminum/cuda/cuda.hpp"
 
 namespace Al {
 namespace internal {
